[spi-v2] (Part 2) Implement adapters to convert SPI versions
FromV1 has been added which will wrap V1 SPI object and into
a V2 object. The pluggable layer will later use this to
convert a V1 plugin to a V2 plugin.

In order for conversion of V1 -> V2 to work as expected, the
reverse is also needed to convert V2 arguments into V1
arguments which is handled by the ToV1.