CLOUD-6963. No throw if allocated instance count < min
When allocating instances, if the count of allocated instances is
less than min count, we are not going to throw for on demand
instances (spot instance is already implemented this way).

In the pipeline job, if allocated instance count < min, we are
only going to terminate pipeline if there is unrecoverable
exception thrown; otherwise we are going to retry.